add Vpolyintersection diameter computation and volume approximation
