+++ include/volume/volume.h
@@ -25,7 +25,7 @@
 #include "zpolytope.h"
 #include "ball.h"
 #include "ballintersectconvex.h"
-#include "vpolyintersectvpoly.h"
+//#include "vpolyintersectvpoly.h"
 #include "samplers.h"
 #include "rounding.h"
 #include "gaussian_samplers.h"
